Replace next with vite
