+++ assets/index.css
@@ -2,14 +2,12 @@
   Rhythm for font sizes:
   Consecutive sizes differ by the cube root of 3 (rounded to the nearest
   integer.) Font size doubles every 3 steps.
-
-
 */
  @import url('https://fonts.googleapis.com/css?family=IBM+Plex+Mono|IBM+Plex+Sans:400,400i,700&display=swap');
 
 body {
   margin: 0;
-  padding: 0;
+  padding: 30px;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
@@ -64,7 +62,6 @@ h5 {
   letter-spacing: 1px;
 }
 
-
 blockquote, pre {
   padding: 0 16px;
   margin: 16px 0;
@@ -83,6 +80,183 @@ a:hover {
   box-shadow: -4px 0 0 0 #f36, 4px 0 0 0 #f36;
   color: #fff;
   text-decoration: none;
+}
+
+
+/** Styles for the example **/
+
+.Example {
+  text-align: left;
+  width: 100%;
+  line-height: 1em;
+}
+
+.Pagination {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-bottom: 2em;
+}
+
+.Pagination > * {
+  font-size: 1.25em;
+  line-height: 1em;
+  padding: 1em 0;
+  background: #eee;
+  color: #333;
+}
+
+.Pagination > span {
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.Pagination > button {
+  width: 3em;
+  border-radius: 0;
+  border: none;
+  cursor: pointer;
+}
+
+.Pagination > button:disabled {
+  color: #ccc;
+  cursor: default;
+}
+
+.Pagination > button:hover:not(:disabled) {
+  background: #fff;
+  box-shadow: inset 0 0 16px 8px #eee;
+}
+
+.Pagination > button.PrevPage {
+  border-radius: 100% 0 0 100%;
+}
+
+.Pagination > button.NextPage {
+  border-radius: 0 100% 100% 0;
+}
+
+.Query {
+  position: fixed;
+  display: block;
+  top: -1px;
+  left: -60%;
+  height: 100%;
+  width: 60%;
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace;
+  font-size: 16px;
+  line-height: 1.5em;
+  box-sizing: content-box;
+  border: none;
+  background: #eee;
+  z-index: 1;
+  padding: 24px;
+  transition: all 0.2s ease-out;
+  cursor: pointer;
+}
+
+.Query:focus {
+  left: 0;
+  cursor: text;
+  background: #fff;
+  box-shadow: 0 0 8px 0 rgba(0, 0, 0, 0.5);
+}
+
+.Spinner {
+  position: fixed;
+  top: 0;
+  right: 0;
+  left: 0;
+  bottom: 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 2em;
+  background: rgba(0, 0, 0, 0.25);
+  color: #fff;
+  text-shadow: 0 0 8px #000;
+}
+
+.Visitor {
+  display: flex;
+  flex-direction: column;
+  padding: 0px;
+  border: 0.5em solid #eee;
+  background: #eee;
+  color: #666;
+  border-radius: 2em;
+  cursor: pointer;
+  overflow: hidden;
+}
+
+.Visitor--muted {
+  opacity: 0.25;
+}
+
+.Visitor:hover:not(:disabled) {
+  background: #fff;
+  box-shadow: inset 0 0 16px 8px #eee;
+}
+
+.Visitor-avatar {
+  flex: 0 0 auto;
+  width: 3em;
+  height: 3em;
+  border-radius: 100%;
+  overflow: hidden;
+}
+
+.Visitor-meta {
+  display: flex;
+  align-items: center;
+  flex: 1 1 0;
+}
+
+.Visitor-name {
+  color: #333;
+  font-size: 1.5em;
+  line-height: 1.333em;
+  padding: 0 0.5em;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.Visitor-ts {
+  text-align: center;
+  color: #999;
+}
+
+.Visitor-pages {
+  flex: 1 1 0;
+  margin: 0.1em 0 1em 1.5em;
+}
+
+.Visitor-page {
+  position: relative;
+  padding-left: 1em;
+}
+
+.Visitor-page::before {
+  content: '';
+  position: absolute;
+  border: 2px solid #999;
+  height: 8px;
+  width: 8px;
+  border-radius: 100%;
+  left: -5px;
+  bottom: 0.1em;
+}
+
+.Visitor-page:last-child::before {
+  background: #999;
+}
+
+.List {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(12em, 1fr));
+  grid-gap: 1em;
 }
