+++ package.json
@@ -12,12 +12,8 @@
     "utest": "jest --testPathIgnorePatterns=example",
     "test": "NODE_ENV=testing jest",
     "package": "./scripts/package",
-    "typedef": "tsc --allowJs --declaration --emitDeclarationOnly",
     "format": "prettier --write '{src,scripts,example,pages}/**/*.{js,css,html,json,md}' && eslint --fix scripts src pages",
-    "next": "NEXTJS=1 next",
-    "build": "NEXTJS=1 next build",
-    "start": "node --es-module-specifier-resolution=node src/website/server.js",
-    "rogue": "node --es-module-specifier-resolution=node src/rogue-admin/server.js"
+    "start": "node src/website/server.js"
   },
   "workspaces": [
     "src/*",
@@ -38,11 +34,8 @@
     "@babel/plugin-transform-runtime": "^7.15.0",
     "@babel/preset-env": "^7.15.0",
     "@babel/preset-react": "^7.10.4",
-    "@mapbox/rehype-prism": "^0.6.0",
-    "@mdx-js/loader": "^1.6.6",
-    "@mdx-js/mdx": "^1.6.6",
-    "@next/mdx": "10.2.3",
     "@testing-library/react-hooks": "^7.0.1",
+    "@vitejs/plugin-react-refresh": "^1.3.6",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^27.0.6",
     "babel-plugin-add-module-exports": "^1.0.2",
@@ -52,17 +45,16 @@
     "eslint-plugin-prettier": "^3.4.1",
     "eslint-plugin-react": "^7.20.3",
     "express": "^4.17.1",
-    "fakataaar": "^1.0.0",
     "faker": "^5.1.0",
     "global": "^4.4.0",
     "globby": "^12.0.1",
+    "graphiql": "^1.4.2",
     "graphql": "^15.5.1",
     "graphql-tag": "^2.12.5",
     "jest": "^27.0.6",
     "lodash": "^4.17.19",
     "merge-async-iterators": "^0.2.1",
     "nanoid": "^3.1.25",
-    "next": "10.2.3",
     "pg": "^8.7.1",
     "prettier": "^2.3.2",
     "prop-types": "^15.7.2",
@@ -74,6 +66,7 @@
     "sql-template-tag": "^4.0.0",
     "text-encoder-lite": "^2.0.0",
     "typescript": "^4.3.5",
+    "vite": "^2.5.0",
     "ws": "^7.0.0",
     "yargs": "^17.1.1"
   },
