+++ src/common/util.js
@@ -1,21 +1,5 @@
-import util from 'util';
-
-const opts = {
-  showHidden: false,
-  depth: 3,
-  colors: false,
-  customInspect: true,
-  showProxy: false,
-  maxArrayLength: 3,
-  maxStringLength: 10,
-  breakLength: Infinity,
-  compact: true,
-  sorted: false,
-  getters: false,
-};
-
 export function err(message, { cause, ...args } = {}) {
-  const e = new Error(message + (args ? ' ' + util.inspect(args, opts) : ''));
+  const e = new Error(message + (args ? ' ' + JSON.stringify(args) : ''));
   e.cause = cause;
   throw e;
 }
