+++ src/react/index.js
@@ -1,4 +1,4 @@
-export { GraffyProvider } from './GraffyContext.js';
+export { GraffyProvider } from './GraffyContext.jsx';
 export { default as Query } from './Query.js';
 export { default as useQuery } from './useQuery.js';
 export { default as useStore } from './useStore.js';
