+++ src/website/components/Example/Spinner.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Spinner() {
+  return <div className="Spinner">Loading...</div>;
+}
