+++ src/website/components/Navigation/index.js
@@ -1,7 +1,11 @@
 import React from 'react';
-import Link from 'next/link';
+import linkMod from 'next/link';
+
+const { default: Link } = linkMod;
 
 function Navigation({ menu }) {
+  console.log('link', Link);
+  // return null;
   return (
     <ul>
       {menu.map(({ title, url, external, children }) => (
@@ -15,7 +19,7 @@ function Navigation({ menu }) {
           {children && <Navigation menu={children} />}
         </li>
       ))}
-      <style jsx>{`
+      {/*<style jsx>{`
         text-decoration: none;
         list-style-type: none;
         ul {
@@ -34,7 +38,7 @@ function Navigation({ menu }) {
           padding: 0 4rem;
           box-shadow: none;
         }
-      `}</style>
+      `}</style>*/}
     </ul>
   );
 }
