+++ src/website/example.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+// import './index.css'
+import Example from './components/Example';
+
+ReactDOM.render(
+  <React.StrictMode>
+    <Example />
+  </React.StrictMode>,
+  document.getElementById('root'),
+);
