+++ src/website/server.js
@@ -1,21 +1,16 @@
-import { join } from 'path';
-import next from 'next';
 import express from 'express';
+import { createServer as createViteServer } from 'vite';
+
 import debug from 'debug';
 import Graffy from '@graffy/core';
 import fill from '@graffy/fill';
 import memory from '@graffy/memory';
 import { httpServer, wsServer } from '@graffy/server';
 import mock from './services/mockVisitorList.js';
-import { fileURLToPath } from 'url';
 
 process.env.NEXTJS = 1; // Use the Next.js babel config.
 
 const log = debug('graffy:website:server');
-const dev = process.env.NODE_ENV !== 'production';
-const dir = join(fileURLToPath(import.meta.url), '..', '..', '..');
-const nextApp = next({ dev, dir });
-const handle = nextApp.getRequestHandler();
 const port = process.env.PORT || 3000;
 
 const store = new Graffy();
@@ -23,24 +18,28 @@ store.use(fill());
 store.use(memory());
 store.use(mock);
 
-nextApp
-  .prepare()
-  .then(() => {
-    const server = express();
+// const express = require('express')
 
-    server.use('/api', httpServer(store));
-    server.get('*', (req, res) => {
-      handle(req, res);
-    });
+async function createServer() {
+  const app = express();
 
-    const conn = server.listen(port, (err) => {
-      if (err) throw err;
-      if (process.send) process.send('ready');
-      log(`Ready on port ${port}`);
-    });
-
-    conn.on('upgrade', wsServer(store));
-  })
-  .catch((e) => {
-    log('NextApp Prepare Error:', e);
+  // Create vite server in middleware mode.
+  const vite = await createViteServer({
+    server: { middlewareMode: 'html' },
   });
+
+  app.use('/api', httpServer(store));
+
+  // Use vite's connect instance as middleware
+  app.use(vite.middlewares);
+
+  const conn = app.listen(port, (err) => {
+    if (err) throw err;
+    if (process.send) process.send('ready');
+    log(`Ready on port ${port}`);
+  });
+
+  conn.on('upgrade', wsServer(store));
+}
+
+createServer();
