+++ src/website/services/mockVisitorList.js
@@ -1,5 +1,4 @@
 import faker from 'faker';
-import fakataaar from 'fakataaar';
 import debug from 'debug';
 const log = debug('graffy:website:server');
 import { encodeGraph } from '@graffy/common';
@@ -68,9 +67,7 @@ function simulate() {
 function visitorInfo() {
   return {
     name: faker.internet.userName(),
-    avatar:
-      'data:image/svg+xml;base64,' +
-      Buffer.from(fakataaar()).toString('base64'),
+    avatar: faker.image.avatar(),
     pageviews: [{ $key: [ts], $val: faker.system.directoryPath() }],
   };
 }
